Enable Dual Compiling for Swift 3 and 4
This library is well aimed at Linux and the Raspberry Pi in particular, but the Raspberry Pi is stuck on Swift 3.1.1 if targeting the Zero.

This change enables it to build/test for both Swift 3 and 4 at the same time. The code remains mostly Swift 4, but uses wrappers around the Swift 3 versions of functions to allow most of the Swift 4 code to continue to operate.

The one unfortunate change is that Swift 3.1.1 doesn’t like the “$0 ? () : nil” syntax in count()’s map. While it technically uses fewer bytes marking the entries with empty tuples, it creates issues resolving types in Swift 3. As a workaround, we can just use a Bool instead.

If I can fix this last issue later a different way, I will.